requirements.txt
colorama==0.4.6
numpy==1.26.2
PyYAML==6.0.1